+++ src/components/traits/i-observe-dom/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2022-??-??)
+
+#### :boom: Breaking Change
+
+* Renamed `filterNodes` to `filterMutations`
+
 ## v3.0.0-rc.184 (2021-05-12)
 
 #### :rocket: New Feature